backpacks and pouches
Med bags storage from 40 to 24, removed a bottle of morphine, a bottle of amino, and a gauze pack. They can now hold gauze and tourniquet satchels.

Added the SSG82 ammo pouch, can hold two magazines. Made it so NVA RPK gunners start with their pouches full of big magazines (4)- do not put any of these in the armory.  Made it so every "ugrenade" type pouch starts with 4 frag and 2 smoke shells- do not put any frag in the armories.